requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langgraph==0.2.4

pinecone-client==3.2.2

sentence-transformers==2.7.0
pypdf==4.2.0

streamlit==1.36.0

python-dotenv==1.0.1

ollama==0.3.3
